Renaming Types, suffixing them with Type.
