+++ lib/Pheasant/Types/BooleanType.php
@@ -5,16 +5,16 @@ namespace Pheasant\Types;
 use \Pheasant\Database\TypedValue;
 
 /**
- * A date and time type that persists to a unix timestamp
+ * A basic string type
  */
-class UnixTimestamp extends Base
+class BooleanType extends BaseType
 {
     /* (non-phpdoc)
      * @see \Pheasant\Type::columnSql
      */
     public function columnSql($column, $platform)
     {
-        return $platform->columnSql($column, 'int', $this->options());
+        return $platform->columnSql($column, "boolean", $this->options());
     }
 
     /* (non-phpdoc)
@@ -22,7 +22,7 @@ class UnixTimestamp extends Base
      */
     public function unmarshal($value)
     {
-        return new \DateTime('@'.$value);
+        return (bool) $value;
     }
 
     /* (non-phpdoc)
@@ -30,6 +30,6 @@ class UnixTimestamp extends Base
      */
     public function marshal($value)
     {
-        return new TypedValue($value->getTimestamp());
+        return new TypedValue((bool)$value);
     }
 }
